Needs updated Memory Functions
Runs 3 jobs to completion now, goes to put the 4th when a 5th comes in
and cannot be given memory.

ISSUE: For some reason job 3 terminates abnormally. It runs to
completion, but calls tro() on completion. Any attempt to put it into
the CPU will be met with an error and it does not appear on the Job
Table afterwards if the program crashes.