Fixed Scheduler issue
When a job with high priority wanted to be blocked, the findjob didn't
have a check to make sure that the first thing on the map wasn't that
thing. Also TRO needed a way to not stall the CPU if there were jobs in
core. Current issue is I/O scheduler